+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add SAEec06 and SAEec09 SLOTs.
 - Add data page and extended data page to builder.
+- Implement `PartialEq` to mask of priority bits when comparing ids.
 
 ### Added